# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  controller.SuccessResponse:
    properties:
      result:
        type: string
    type: object
...
  domain.AccountResponses:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      explanatory_text:
        type: string
      frameworks:
        items:
          $ref: '#/definitions/repository.Framework'
        type: array
      icon:
        type: string
      locate:
        type: string
      rate:
        type: integer
      show_locate:
        type: boolean
      show_rate:
        type: boolean
      tech_tags:
        items:
          $ref: '#/definitions/repository.TechTag'
        type: array
      username:
        type: string
    type: object
...
  domain.CreateAccountRequest:
    properties:
      account_id:
        type: string
      explanatory_text:
        type: string
      frameworks:
        type: string
      locate_id:
        type: integer
      show_locate:
        type: boolean
      show_rate:
        type: boolean
      tech_tags:
        type: string
      username:
        type: string
    required:
    - account_id
    - locate_id
    - username
    type: object
  domain.CreateFollowRequestBody:
    properties:
      to_account_id:
        type: string
    required:
    - to_account_id
    type: object
...
  domain.FollowResponse:
    properties:
      account_id:
        type: string
      icon:
        type: string
      username:
        type: string
    type: object
...
  domain.UpdateAccountRequest:
    properties:
      explanatory_text:
        type: string
      frameworks:
        type: string
      locate_id:
        type: integer
      show_locate:
        type: boolean
      show_rate:
        type: boolean
      tech_tags:
        type: string
      username:
        type: string
    type: object
...
  repository.Follow:
    properties:
      create_at:
        type: string
      from_account_id:
        type: string
      to_account_id:
        type: string
    type: object
  repository.Framework:
    properties:
      framework:
        type: string
      framework_id:
        type: integer
      tech_tag_id:
        type: integer
    type: object
  repository.Locate:
    properties:
      locate_id:
        type: integer
      name:
        type: string
    type: object
  repository.StatusTag:
    properties:
      status:
        type: string
      status_id:
        type: integer
    type: object
  repository.TechTag:
    properties:
      icon:
        type: string
      language:
        type: string
      tech_tag_id:
        type: integer
    type: object
host: https://seaffood.com
info:
  contact:
    email: oogiriminister@gmail.com
    name: murasame
    url: https://twitter.com/fresh_salmon256
  description: HackPortal Backend API serice
  license:
    name: No-license
    url: No-license
  termsOfService: https://seaffood.com/api
  title: Hack Hack Backend API
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - multipart/form-data
      description: Create an account from the requested body
      parameters:
      - description: Create Account Request
        in: body
        name: CreateAccountRequest
        required: true
        schema:
          $ref: '#/definitions/domain.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: create success response
          schema:
            $ref: '#/definitions/domain.AccountResponses'
        "400":
          description: bad request response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: server error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create new account
      tags:
      - Accounts
  /accounts/{account_id}:
    delete:
      description: Only you can delete your account (logical delete)
      parameters:
      - description: Accounts API wildcard
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: delete success response
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: bad request response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: server error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Remove Account
      tags:
      - Accounts
    get:
      description: Return a account from the id specified in the path
      parameters:
      - description: Accounts API wildcard
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get success response
          schema:
            $ref: '#/definitions/domain.AccountResponses'
        "400":
          description: bad request response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: server error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get account
      tags:
      - Accounts
    put:
      description: Update account info from requested body
      parameters:
      - description: Accounts API wildcard
        in: path
        name: account_id
        required: true
        type: string
      - description: Update Account Request Body
        in: body
        name: UpdateAccountRequest
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update success response
          schema:
            $ref: '#/definitions/domain.AccountResponses'
        "400":
          description: bad request response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: server error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Update Account
      tags:
      - Accounts
...
  /accounts/{from_account_id}/follow:
    delete:
      description: Remove follow account
      parameters:
      - description: Accounts API wildcard
        in: path
        name: from_account_id
        required: true
        type: string
      - in: query
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Remove follow
      tags:
      - Accounts
    get:
      description: Get follow account
      parameters:
      - description: Accounts API wildcard
        in: path
        name: from_account_id
        required: true
        type: string
      - in: query
        name: mode
        type: boolean
      - in: query
        name: page_id
        type: string
      - in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/domain.FollowResponse'
            type: array
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get follow
      tags:
      - Accounts
    post:
      description: Follow!!!!!!!!
      parameters:
      - description: Accounts API wildcard
        in: path
        name: from_account_id
        required: true
        type: string
      - description: create Follow Request Body
        in: body
        name: CreateFollowRequestBody
        required: true
        schema:
          $ref: '#/definitions/domain.CreateFollowRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/repository.Follow'
            type: array
        "400":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create Follow
      tags:
      - Accounts
  /frameworks:
    get:
      description: Get Frameworks
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/repository.Framework'
            type: array
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get Frameworks
      tags:
      - Frameworks
...
  /locates:
    get:
      description: Get Frameworks
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/repository.Locate'
            type: array
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get Frameworks
      tags:
      - Locates
...
  /status_tags:
    get:
      description: Get Frameworks
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/repository.StatusTag'
            type: array
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get Frameworks
      tags:
      - TechTags
  /tech_tags:
    get:
      description: Get Frameworks
      produces:
      - application/json
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: '#/definitions/repository.TechTag'
            type: array
        "500":
          description: error response
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get Frameworks
      tags:
      - TechTags
swagger: "2.0"